MD5.xs: silence possible loss of data warnings on Win32
When building Perl on Win32 we see warnings like this:

    MD5.xs(723) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(723) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(723) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(724) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(724) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(724) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data
    MD5.xs(725) : warning C4244: '=' : conversion from 'U32' to 'char', possible loss of data

We would like Perl to build without warnings on all platforms so we
would like to get these warnings silenced. You don't see them on gcc
builds as gcc has different ideas about what to warn for in these cases.
You can enable -Wconversion but it complains about a ton of stuff that
VC wouldn't. Generally VC warnings are at least plausibly mistakes,
the -Wconversion ones are mostly things that are not. :-(